Do not init anything at all in the Stripes' process
