+++ WooCommerce/src/debug/kotlin/com.woocommerce.android/WooCommerceDebug.kt
@@ -1,6 +1,5 @@
 package com.woocommerce.android
 
-import android.os.Build
 import android.os.StrictMode
 import android.os.StrictMode.ThreadPolicy
 import android.os.StrictMode.VmPolicy
@@ -68,7 +67,7 @@ class WooCommerceDebug : WooCommerce() {
 
     /**
      * Tap 2 pay Stripe library uses webview, and apparently they also enable debugging of it (at least in the current
-     * beta version). This method changes dirrectory where logs are stored, otherwise it crashes with
+     * beta version). This method changes directory where logs are stored, otherwise it crashes with
      * Caused by: java.lang.RuntimeException: Using WebView from more than one process at once with the same data
      * directory is not supported. https://crbug.com/558377 :
      * Current process com.stripe.cots.aidlservice (pid 7378), lock owner com.woocommerce.android (pid 6427)
@@ -78,7 +77,7 @@ class WooCommerceDebug : WooCommerce() {
      * > since the data directories must be distinct.
      */
     private fun enableWebContentDebugging() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+        if (SystemVersionUtils.isAtLeastP()) {
             val process = getProcessName()
             if (packageName != process) WebView.setDataDirectorySuffix(process)
         }
