+++ WooCommerce/src/main/kotlin/com/woocommerce/android/WooCommerce.kt
@@ -1,15 +1,11 @@
 package com.woocommerce.android
 
-import android.app.ActivityManager
 import android.app.Application
-import android.content.Context
-import android.os.Build
 import androidx.hilt.work.HiltWorkerFactory
 import androidx.work.Configuration
 import com.android.volley.VolleyLog
-import com.google.firebase.FirebaseApp
 import com.woocommerce.android.config.RemoteConfigRepository
-import com.woocommerce.android.ui.payments.taptopay.IsTapToPayAvailable
+import com.woocommerce.android.extensions.getCurrentProcessName
 import com.yarolegovich.wellsql.WellSql
 import dagger.Lazy
 import dagger.android.AndroidInjector
@@ -18,8 +14,7 @@ import dagger.android.HasAndroidInjector
 import javax.inject.Inject
 
 open class WooCommerce : Application(), HasAndroidInjector, Configuration.Provider {
-    @Inject
-    lateinit var androidInjector: DispatchingAndroidInjector<Any>
+    @Inject lateinit var androidInjector: DispatchingAndroidInjector<Any>
 
     // inject it lazily to avoid creating it before initializing WellSql
     @Inject lateinit var appInitializer: Lazy<AppInitializer>
@@ -28,12 +23,14 @@ open class WooCommerce : Application(), HasAndroidInjector, Configuration.Provid
 
     @Inject lateinit var workerFactory: HiltWorkerFactory
 
-    @Inject lateinit var isTapToPayAvailable: IsTapToPayAvailable
-
     override fun onCreate() {
         super.onCreate()
 
-        initFirebaseIfSeparateProcess()
+        // Stripe Tap to Pay library starts it's own process. That causes the crash:
+        //  > Caused by: java.lang.IllegalStateException: Default FirebaseApp is not initialized in this process
+        //  > com.stripe.cots.aidlservice. Make sure to call FirebaseApp.initializeApp(Context) first.
+        // In this case we don't want to initialize any Firebase (or any at all) features of the app in their process.
+        if (getCurrentProcessName() == TAP_TO_PAY_STRIPE_PROCESS_NAME) return
 
         // Disables Volley debug logging on release build and prevents the "Marker added to finished log" crash
         // https://github.com/woocommerce/woocommerce-android/issues/817
@@ -56,30 +53,6 @@ open class WooCommerce : Application(), HasAndroidInjector, Configuration.Provid
     }
 
     override fun androidInjector(): AndroidInjector<Any> = androidInjector
-
-    /**
-     * Looks like that stripe Tap to Pay starts it's own process
-     * > Caused by: java.lang.IllegalStateException: Default FirebaseApp is not initialized in this process
-     * > com.stripe.cots.aidlservice. Make sure to call FirebaseApp.initializeApp(Context) first.
-     *
-     * In this case we have to initialise Firebase manually in this process as per documentation
-     * https://firebase.google.com/docs/reference/android/com/google/firebase/FirebaseApp
-     */
-    private fun initFirebaseIfSeparateProcess() {
-        if (isTapToPayAvailable()) {
-            if (getCurrentProcessName() == TAP_TO_PAY_STRIPE_PROCESS_NAME) {
-                FirebaseApp.initializeApp(this)
-            }
-        }
-    }
-
-    private fun getCurrentProcessName() =
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            getProcessName()
-        } else {
-            val am = getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
-            am.runningAppProcesses.firstOrNull { it.pid == android.os.Process.myPid() }?.processName
-        }
 
     companion object {
         private const val TAP_TO_PAY_STRIPE_PROCESS_NAME = "com.stripe.cots.aidlservice"
