+++ WooCommerce/src/main/kotlin/com/woocommerce/android/extensions/ContextExt.kt
@@ -1,7 +1,26 @@
 package com.woocommerce.android.extensions
 
+import android.app.ActivityManager
+import android.app.Application
+import android.content.ClipData
+import android.content.ClipboardManager
 import android.content.Context
 import androidx.annotation.ColorRes
 import androidx.core.content.ContextCompat
+import com.woocommerce.android.util.SystemVersionUtils
 
 fun Context.getColorCompat(@ColorRes colorRes: Int) = ContextCompat.getColor(this, colorRes)
+
+fun Context.copyToClipboard(label: String, text: String) {
+    with(getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager) {
+        setPrimaryClip(ClipData.newPlainText(label, text))
+    }
+}
+
+fun Context.getCurrentProcessName() =
+    if (SystemVersionUtils.isAtLeastP()) {
+        Application.getProcessName()
+    } else {
+        val am = getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
+        am.runningAppProcesses.firstOrNull { it.pid == android.os.Process.myPid() }?.processName
+    }
