+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/coupons/details/CouponDetailsFragment.kt
@@ -10,6 +10,7 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.woocommerce.android.R
 import com.woocommerce.android.databinding.FragmentCouponDetailsBinding
+import com.woocommerce.android.extensions.copyToClipboard
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.base.UIMessageResolver
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
@@ -18,7 +19,6 @@ import com.woocommerce.android.ui.coupons.details.CouponDetailsViewModel.ShareCo
 import com.woocommerce.android.ui.coupons.details.CouponDetailsViewModel.ShowEditCoupon
 import com.woocommerce.android.ui.main.AppBarStatus
 import com.woocommerce.android.util.WooLog
-import com.woocommerce.android.util.copyToClipboard
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowSnackbar
 import dagger.hilt.android.AndroidEntryPoint
